feat(downloader): add parallel download logic and TNAFlix integration
### Details:
- **Parallel Downloading:**
  - Introduced `download_parallel` and `download_parallel_resume` methods in `HttpDownloader` for fine-grained parallel downloading using range requests.
  - Added `ProgressGuard` for managing and safely aborting progress reporting tasks.
  - Implemented chunk merging (`merge_chunks`) and appending logic (`append_chunks`) with automatic cleanup of temporary chunk files.
  - Improved download concurrency and progress tracking using `tokio` utilities and atomic counters.

- **TNAFlix Integration:**
  - Added modules for extracting video sources and metadata from TNAFlix network sites, including support for `parse_video_sources`, `extract_config_url`, and multi-pattern fallback logic.
  - Provided utilities for parsing HTML and XML formats, handling CDN URLs, and extracting video metadata (e.g., resolution, format type).

- **JSON-LD Support:**
  - Implemented JSON-LD parsing (`json_ld.rs`) for structured metadata extraction (e.g., video title, tags, categories, interaction statistics).
  - Added support for parsing thumbnails, video statistics, and genres directly from HTML script tags.

- **Reusable Utilities and Configurations:**
  - Introduced `DownloaderConfig` for consolidated configuration management (e.g., buffer size, retry strategy, concurrent fragments).
  - Added reusable chunk calculation logic for determining chunk sizes and batch counts.

### Impact:
This commit significantly enhances the downloader's functionality by introducing parallelism, robust resumption capabilities, and extensive support for TNAFlix's video extraction pipelines with JSON-LD integration.

### Notes:
- Comprehensive testing added for parallel downloading, chunk merging, TNAFlix extraction logic, and JSON-LD parsing.
- Improved modularity with clear separation of downloader, extractor, and utility concerns.
