+++ crates/rdlp-downloader/src/http/mod.rs
@@ -1,93 +1,29 @@
-use async_trait::async_trait;
-use futures::StreamExt;
-use log::{debug, error, info, warn};
-use rdlp_core::{check_http_response, retry_with_backoff, DownloadProgress, DownloadStats, Downloader, ProgressCallback, Result, RetryConfig, RdlpError};
-use std::path::Path;
-use std::sync::Arc;
-use std::sync::atomic::AtomicU64;
-use std::time::{Duration, Instant};
-use tokio::fs::File;
-use tokio::io::{AsyncWriteExt, BufWriter};
-use crate::chunking::{calculate_chunks, ChunkSizeStrategy};
+//! HTTP/HTTPS downloader implementation
+//!
+//! Provides HTTP downloading with parallel chunk support, resume capability,
+//! and automatic retry logic.
+
+mod config;
+mod parallel;
 
 #[cfg(test)]
 mod tests;
 
-/// Global atomic counter for generating unique download IDs
-/// This prevents chunk file collisions when multiple downloads run concurrently
-static DOWNLOAD_ID_COUNTER: AtomicU64 = AtomicU64::new(0);
+use async_trait::async_trait;
+use futures::StreamExt;
+use log::{debug, info, warn};
+use rdlp_core::{
+    check_http_response, retry_with_backoff, DownloadProgress, DownloadStats, Downloader,
+    ProgressCallback, Result, RetryConfig, RdlpError,
+};
+use std::path::Path;
+use std::sync::Arc;
+use std::time::{Duration, Instant};
+use tokio::fs::File;
+use tokio::io::{AsyncWriteExt, BufWriter};
 
-/// RAII guard for progress reporter tasks
-///
-/// Ensures the progress reporter task is aborted when the guard goes out of scope,
-/// preventing task leaks on early returns or errors.
-struct ProgressGuard(Option<tokio::task::JoinHandle<()>>);
-
-impl ProgressGuard {
-    fn new(task: Option<tokio::task::JoinHandle<()>>) -> Self {
-        Self(task)
-    }
-}
-
-impl Drop for ProgressGuard {
-    fn drop(&mut self) {
-        if let Some(task) = self.0.take() {
-            task.abort();
-        }
-    }
-}
-
-/// Cleanup chunk files on error
-async fn cleanup_chunk_files(temp_dir: &Path, filename: &str, download_id: u64, total_chunks: usize) {
-    info!("Cleaning up {total_chunks} partial chunk files...");
-    let mut deleted = 0;
-    for chunk_id in 0..total_chunks {
-        let chunk_path = temp_dir.join(format!("{filename}.{download_id}.part{chunk_id}"));
-        if tokio::fs::remove_file(&chunk_path).await.is_ok() {
-            deleted += 1;
-        }
-    }
-    debug!("Deleted {deleted} chunk files");
-}
-
-/// Downloader configuration (shared across clones via Arc)
-///
-/// This struct consolidates all config fields into a single Arc,
-/// making HttpDownloader clones truly zero-cost (~5ns Arc clone vs ~24 bytes field copies).
-///
-/// **Memory optimization:**
-/// - Before: 591 clones × 24 bytes = ~14 KB copied
-/// - After: 591 Arc clones × 8 bytes = ~5 KB pointers
-/// - **Savings: ~9 KB per download**
-#[derive(Clone)]
-struct DownloaderConfig {
-    buffer_size: usize,
-    retry_config: RetryConfig,
-    concurrent_fragments: usize,
-    chunk_strategy: ChunkSizeStrategy,
-}
-
-impl Default for DownloaderConfig {
-    fn default() -> Self {
-        // Calculate optimal concurrent connections based on CPU threads
-        // For I/O-bound workloads like HTTP downloads:
-        // - Tokio can handle many more tasks than CPU cores
-        // - Research shows: aria2 uses 4-16 connections, yt-dlp defaults to 1
-        // - Formula: min(available_parallelism, 8) for balanced I/O saturation
-        //   * Too few: underutilizes bandwidth
-        //   * Too many: connection overhead, server rate limiting
-        let concurrent_fragments = std::thread::available_parallelism()
-            .map(|n| n.get().min(8))
-            .unwrap_or(4); // Fallback to 4 if detection fails
-
-        Self {
-            buffer_size: 8192, // 8KB buffer
-            retry_config: RetryConfig::default_config(),
-            concurrent_fragments,
-            chunk_strategy: ChunkSizeStrategy::Auto, // Use automatic power-of-two chunking
-        }
-    }
-}
+use crate::chunking::ChunkSizeStrategy;
+use config::DownloaderConfig;
 
 /// HTTP/HTTPS downloader
 ///
@@ -95,7 +31,7 @@ impl Default for DownloaderConfig {
 #[derive(Clone)]
 pub struct HttpDownloader {
     client: reqwest::Client,
-    config: Arc<DownloaderConfig>,
+    pub(crate) config: Arc<DownloaderConfig>,
 }
 
 impl HttpDownloader {
@@ -134,7 +70,7 @@ impl HttpDownloader {
     /// Set number of concurrent fragment downloads
     #[must_use = "builder methods consume self and return a new instance"]
     pub fn with_concurrent_fragments(mut self, count: usize) -> Self {
-        Arc::make_mut(&mut self.config).concurrent_fragments = count.max(1); // At least 1
+        Arc::make_mut(&mut self.config).concurrent_fragments = count.max(1);
         self
     }
 
@@ -147,20 +83,23 @@ impl HttpDownloader {
 
     /// Check if server supports range requests
     async fn supports_ranges(&self, url: &str) -> Result<bool> {
-        let response = retry_with_backoff(&self.config.retry_config, "HTTP HEAD (range check)", |_attempt| {
-            let client = self.client.clone();
-            let url = url.to_string();
-            async move {
-                client
-                    .head(&url)
-                    .send()
-                    .await
-                    .map_err(|e| RdlpError::Network(format!("Failed to HEAD request: {e}")))
-            }
-        })
+        let response = retry_with_backoff(
+            &self.config.retry_config,
+            "HTTP HEAD (range check)",
+            |_attempt| {
+                let client = self.client.clone();
+                let url = url.to_string();
+                async move {
+                    client
+                        .head(&url)
+                        .send()
+                        .await
+                        .map_err(|e| RdlpError::Network(format!("Failed to HEAD request: {e}")))
+                }
+            },
+        )
         .await?;
 
-        // Check for Accept-Ranges header
         Ok(response
             .headers()
             .get("accept-ranges")
@@ -170,65 +109,35 @@ impl HttpDownloader {
     }
 
     /// Download a specific byte range with shared progress tracking
-    ///
-    /// # Arc<AtomicU64> Pattern for Lock-Free Concurrency
-    ///
-    /// The `progress_counter` uses `Arc<AtomicU64>` for efficient shared state:
-    ///
-    /// **Memory Model:**
-    /// ```rust,ignore
-    /// Arc<AtomicU64>
-    /// │
-    /// ├─ Arc: Atomic Reference Counting
-    /// │  └─ Enables cheap cloning across parallel tasks (~5ns per clone)
-    /// │  └─ Thread-safe sharing via atomic refcount
-    /// │
-    /// └─ AtomicU64: Lock-free atomic counter
-    ///    └─ No Mutex needed (no lock contention)
-    ///    └─ Updates via CPU atomic instructions (~2ns vs 50ns for Mutex)
-    ///    └─ Ordering::Relaxed sufficient (no inter-thread synchronization needed)
-    /// ```
-    ///
-    /// **Why not Arc<Mutex<u64>>?** (Phase 1 optimization)
-    /// - Mutex: Async lock acquisition (~50ns) + contention overhead
-    /// - AtomicU64: Single CPU instruction (~2ns) + zero contention
-    /// - **25x faster** for simple counter increments
-    ///
-    /// **Hot path performance:**
-    /// - Called hundreds to thousands of times per download
-    /// - Phase 1 benchmark: 5,400 atomic updates with zero lock contention
-    /// - Savings: ~48ns × 5,400 = 259μs per parallel download
-    ///
-    /// **Thread Safety:**
-    /// - Each parallel chunk task holds an `Arc` clone
-    /// - All chunks increment the same atomic counter concurrently
-    /// - No locks, no contention, no await overhead
-    async fn download_range_with_progress(
+    pub(crate) async fn download_range_with_progress(
         &self,
         url: &str,
         start: u64,
         end: u64,
         chunk_path: &Path,
-        progress_counter: Option<std::sync::Arc<std::sync::atomic::AtomicU64>>,
+        progress_counter: Option<Arc<std::sync::atomic::AtomicU64>>,
     ) -> Result<u64> {
-        let response = retry_with_backoff(&self.config.retry_config, "HTTP GET (range)", |_attempt| {
-            let client = self.client.clone();
-            let url = url.to_string();
-            async move {
-                let response = client
-                    .get(&url)
-                    .header("Range", format!("bytes={start}-{end}"))
-                    .send()
-                    .await
-                    .map_err(|e| RdlpError::Network(format!("Failed to fetch range: {e}")))?;
+        let response = retry_with_backoff(
+            &self.config.retry_config,
+            "HTTP GET (range)",
+            |_attempt| {
+                let client = self.client.clone();
+                let url = url.to_string();
+                async move {
+                    let response = client
+                        .get(&url)
+                        .header("Range", format!("bytes={start}-{end}"))
+                        .send()
+                        .await
+                        .map_err(|e| RdlpError::Network(format!("Failed to fetch range: {e}")))?;
 
-                check_http_response(&response)?;
-                Ok(response)
-            }
-        })
+                    check_http_response(&response)?;
+                    Ok(response)
+                }
+            },
+        )
         .await?;
 
-        // Write chunk to temporary file
         let file = File::create(chunk_path).await.map_err(RdlpError::Io)?;
         let mut writer = BufWriter::with_capacity(self.config.buffer_size, file);
 
@@ -242,24 +151,18 @@ impl HttpDownloader {
             writer.write_all(&chunk).await.map_err(RdlpError::Io)?;
             downloaded += chunk.len() as u64;
 
-            // Update shared progress counter in real-time (lock-free atomic operation)
             if let Some(ref counter) = progress_counter {
-                counter.fetch_add(chunk.len() as u64, std::sync::atomic::Ordering::Relaxed);
+                counter.fetch_add(
+                    chunk.len() as u64,
+                    std::sync::atomic::Ordering::Relaxed,
+                );
             }
         }
 
         writer.flush().await.map_err(RdlpError::Io)?;
         Ok(downloaded)
     }
-}
 
-impl Default for HttpDownloader {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl HttpDownloader {
     /// Sequential download (original method)
     async fn download_sequential(
         &self,
@@ -269,7 +172,6 @@ impl HttpDownloader {
     ) -> Result<DownloadStats> {
         let start_time = Instant::now();
 
-        // Make the request with retry logic
         let response = retry_with_backoff(&self.config.retry_config, "HTTP GET", |_attempt| {
             let client = self.client.clone();
             let url = url.to_string();
@@ -280,40 +182,28 @@ impl HttpDownloader {
                     .await
                     .map_err(|e| RdlpError::Network(format!("Failed to fetch URL: {e}")))?;
 
-                // Check status
                 check_http_response(&response)?;
-
                 Ok(response)
             }
         })
         .await?;
 
-        // Get content length if available
         let total_size = response.content_length();
-
-        // Create output file with buffered writer
-        let file = File::create(path)
-            .await
-            .map_err(RdlpError::Io)?;
+        let file = File::create(path).await.map_err(RdlpError::Io)?;
         let mut writer = BufWriter::with_capacity(self.config.buffer_size, file);
 
-        // Download with progress tracking
         let mut stream = response.bytes_stream();
         let mut downloaded: u64 = 0;
         let mut last_update = Instant::now();
-        let update_interval = Duration::from_millis(100); // Update every 100ms
+        let update_interval = Duration::from_millis(100);
 
         while let Some(chunk_result) = stream.next().await {
             let chunk = chunk_result
                 .map_err(|e| RdlpError::Network(format!("Failed to read chunk: {e}")))?;
 
-            writer.write_all(&chunk)
-                .await
-                .map_err(RdlpError::Io)?;
-
+            writer.write_all(&chunk).await.map_err(RdlpError::Io)?;
             downloaded += chunk.len() as u64;
 
-            // Update progress
             if let Some(ref callback) = progress {
                 let now = Instant::now();
                 if now.duration_since(last_update) >= update_interval {
@@ -331,379 +221,22 @@ impl HttpDownloader {
             }
         }
 
-        // Ensure all data is written to disk
         writer.flush().await.map_err(RdlpError::Io)?;
 
         let duration = start_time.elapsed();
         let stats = DownloadStats::new(downloaded, duration, 0);
 
-        // Notify completion
         if let Some(callback) = progress {
             callback.on_complete(&stats);
         }
 
         Ok(stats)
     }
+}
 
-    /// Parallel download using multiple range requests with fine-grained chunking
-    ///
-    /// Uses power-of-two chunk sizes for optimal performance and `buffer_unordered`
-    /// for automatic batch processing with concurrency limiting.
-    async fn download_parallel(
-        &self,
-        url: &str,
-        path: &Path,
-        total_size: u64,
-        progress: Option<Box<dyn ProgressCallback>>,
-    ) -> Result<DownloadStats> {
-        use std::sync::Arc;
-        use std::sync::atomic::{AtomicU64, Ordering};
-        use futures::stream::{self, StreamExt, TryStreamExt};
-
-        let start_time = Instant::now();
-
-        // Generate unique download ID to prevent chunk file collisions
-        let download_id = DOWNLOAD_ID_COUNTER.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-
-        // Calculate optimal power-of-two chunk sizes
-        let (chunk_size, total_chunks) = calculate_chunks(total_size, self.config.chunk_strategy);
-
-        debug!("Chunk analysis: download_id={download_id}, size={} MB, chunk_size={} KB, chunks={total_chunks}, concurrent={}, batches={}",
-            total_size / 1024 / 1024,
-            chunk_size / 1024,
-            self.config.concurrent_fragments,
-            total_chunks.div_ceil(self.config.concurrent_fragments));
-
-        // Create temporary directory for chunks
-        let temp_dir = path.parent().unwrap_or_else(|| Path::new("."));
-        let filename = path.file_name()
-            .ok_or_else(|| RdlpError::Download("Invalid output path: no filename".to_string()))?
-            .to_string_lossy()
-            .to_string();
-
-        // Shared progress counter (lock-free atomic for better performance)
-        let downloaded = Arc::new(AtomicU64::new(0));
-
-        // Progress reporter task (wrapped in guard for automatic cleanup)
-        let _progress_guard = ProgressGuard::new(if let Some(callback) = progress {
-            let downloaded_clone = downloaded.clone();
-            let start_time_clone = start_time;
-            Some(tokio::spawn(async move {
-                let mut last_update = Instant::now();
-                let update_interval = Duration::from_millis(100);
-
-                loop {
-                    tokio::time::sleep(Duration::from_millis(100)).await;
-                    let now = Instant::now();
-                    if now.duration_since(last_update) >= update_interval {
-                        let bytes_downloaded = downloaded_clone.load(Ordering::Relaxed);
-                        let elapsed = now.duration_since(start_time_clone).as_secs_f64();
-                        let speed = if elapsed > 0.0 {
-                            bytes_downloaded as f64 / elapsed
-                        } else {
-                            0.0
-                        };
-
-                        let progress_info = DownloadProgress::new(bytes_downloaded, Some(total_size), speed);
-                        callback.on_progress(&progress_info);
-                        last_update = now;
-
-                        // Exit when complete
-                        if bytes_downloaded >= total_size {
-                            break;
-                        }
-                    }
-                }
-            }))
-        } else {
-            None
-        });
-
-        info!("Starting parallel download with {} concurrent connections", self.config.concurrent_fragments);
-
-        // Arc<str> optimization: Share URL across all chunks (saves ~59 KB for 591 chunks)
-        let url_shared: Arc<str> = Arc::from(url);
-
-        // Download all chunks using buffer_unordered for automatic batch processing
-        let chunk_results: Vec<u64> = stream::iter(0..total_chunks)
-            .map(|chunk_id| {
-                let start = chunk_id as u64 * chunk_size as u64;
-                let end = if chunk_id == total_chunks - 1 {
-                    total_size - 1 // Last chunk gets remainder
-                } else {
-                    start + chunk_size as u64 - 1
-                };
-
-                let chunk_path = temp_dir.join(format!("{}.{}.part{}",
-                    &filename, download_id, chunk_id));
-
-                // Clone what we need for the async block (O(1) Arc clone)
-                let downloader = Self {
-                    client: self.client.clone(),
-                    config: Arc::clone(&self.config),
-                };
-                let url = Arc::clone(&url_shared);  // Cheap Arc clone (~5ns)
-                let progress_counter = Some(downloaded.clone());
-
-                async move {
-                    let result = downloader.download_range_with_progress(
-                        &url, start, end, &chunk_path, progress_counter
-                    ).await;
-                    if let Err(ref e) = result {
-                        error!("Chunk {chunk_id} failed: {e}");
-                    }
-                    result.map(|bytes| (chunk_id, bytes, chunk_path))
-                }
-            })
-            .buffer_unordered(self.config.concurrent_fragments) // Automatic batch processing!
-            .try_collect::<Vec<_>>()
-            .await
-            .inspect_err(|e| {
-                error!("Download failed: {e}");
-                // Note: Progress reporter is automatically cleaned up by _progress_guard on drop
-                // Cleanup partial chunk files
-                let temp_dir = temp_dir.to_path_buf();
-                let filename = filename.clone();
-                tokio::spawn(async move {
-                    cleanup_chunk_files(&temp_dir, &filename, download_id, total_chunks).await;
-                });
-            })?
-            .into_iter()
-            .map(|(_, bytes, _)| bytes)
-            .collect();
-
-        // Sum up all downloaded bytes
-        let total_downloaded: u64 = chunk_results.iter().sum();
-        info!("All {} chunks completed ({} MB total)", total_chunks, total_downloaded / 1024 / 1024);
-
-        // Note: Progress reporter is automatically cleaned up by _progress_guard on drop
-
-        // Merge chunks into final file in correct order
-        info!("Merging {total_chunks} chunks into final file...");
-        let final_file = File::create(path).await.map_err(RdlpError::Io)?;
-        let mut writer = BufWriter::with_capacity(self.config.buffer_size, final_file);
-
-        let mut deleted_chunks = 0;
-        for chunk_id in 0..total_chunks {
-            let chunk_path = temp_dir.join(format!("{}.{}.part{}",
-                &filename, download_id, chunk_id));
-
-            let mut chunk_file = File::open(&chunk_path).await.map_err(RdlpError::Io)?;
-            tokio::io::copy(&mut chunk_file, &mut writer)
-                .await
-                .map_err(RdlpError::Io)?;
-
-            // Delete chunk file after merging
-            if tokio::fs::remove_file(&chunk_path).await.is_ok() {
-                deleted_chunks += 1;
-            }
-
-            if (chunk_id + 1) % 100 == 0 || chunk_id == total_chunks - 1 {
-                debug!("Merged {}/{} chunks", chunk_id + 1, total_chunks);
-            }
-        }
-        debug!("Cleaned up {deleted_chunks} chunk files");
-
-        writer.flush().await.map_err(RdlpError::Io)?;
-
-        let duration = start_time.elapsed();
-        // Fix Issue #2: Use 0 for retry_count (we don't track retries at this level)
-        let stats = DownloadStats::new(total_downloaded, duration, 0);
-
-        info!("Download complete: {} MB in {:.1}s ({:.1} MB/s)",
-            total_downloaded / 1024 / 1024,
-            duration.as_secs_f64(),
-            (total_downloaded as f64 / duration.as_secs_f64()) / 1024.0 / 1024.0);
-
-        Ok(stats)
-    }
-
-    /// Parallel resume: downloads remaining chunks in parallel and appends to existing file
-    ///
-    /// This method keeps the already-downloaded portion and parallelizes the remaining download
-    /// using power-of-two chunking and `buffer_unordered` for batch processing.
-    async fn download_parallel_resume(
-        &self,
-        url: &str,
-        path: &Path,
-        resume_from: u64,
-        total_size: u64,
-        progress: Option<Box<dyn ProgressCallback>>,
-    ) -> Result<DownloadStats> {
-        use std::sync::Arc;
-        use std::sync::atomic::{AtomicU64, Ordering};
-        use futures::stream::{self, StreamExt, TryStreamExt};
-
-        let start_time = Instant::now();
-        let remaining_size = total_size - resume_from;
-
-        // Generate unique download ID to prevent chunk file collisions
-        let download_id = DOWNLOAD_ID_COUNTER.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-
-        // Calculate optimal power-of-two chunk sizes for remaining data
-        let (chunk_size, total_chunks) = calculate_chunks(remaining_size, self.config.chunk_strategy);
-
-        debug!("Parallel resume: download_id={download_id}, already={} MB ({:.1}%), remaining={} MB, chunk_size={} KB, chunks={total_chunks}, concurrent={}",
-            resume_from / 1024 / 1024,
-            (resume_from as f64 / total_size as f64) * 100.0,
-            remaining_size / 1024 / 1024,
-            chunk_size / 1024,
-            self.config.concurrent_fragments);
-
-        // Create temporary directory for chunks
-        let temp_dir = path.parent().unwrap_or_else(|| Path::new("."));
-        let filename = path.file_name()
-            .ok_or_else(|| RdlpError::Download("Invalid output path: no filename".to_string()))?
-            .to_string_lossy()
-            .to_string();
-
-        // Shared progress counter (starts at resume_from, lock-free atomic)
-        let downloaded = Arc::new(AtomicU64::new(resume_from));
-
-        // Progress reporter task (shows total including already downloaded)
-        let progress_task = if let Some(callback) = progress {
-            let downloaded_clone = downloaded.clone();
-            let start_time_clone = start_time;
-            Some(tokio::spawn(async move {
-                let mut last_update = Instant::now();
-                let update_interval = Duration::from_millis(100);
-
-                loop {
-                    tokio::time::sleep(Duration::from_millis(100)).await;
-                    let now = Instant::now();
-                    if now.duration_since(last_update) >= update_interval {
-                        let bytes_downloaded = downloaded_clone.load(Ordering::Relaxed);
-                        let elapsed = now.duration_since(start_time_clone).as_secs_f64();
-                        let speed = if elapsed > 0.0 {
-                            (bytes_downloaded - resume_from) as f64 / elapsed
-                        } else {
-                            0.0
-                        };
-
-                        let progress_info = DownloadProgress::new(bytes_downloaded, Some(total_size), speed);
-                        callback.on_progress(&progress_info);
-                        last_update = now;
-
-                        // Exit when complete
-                        if bytes_downloaded >= total_size {
-                            break;
-                        }
-                    }
-                }
-            }))
-        } else {
-            None
-        };
-
-        info!("Starting parallel download with {} concurrent connections", self.config.concurrent_fragments);
-
-        // Arc<str> optimization: Share URL across all chunks (saves ~59 KB for 591 chunks)
-        let url_shared: Arc<str> = Arc::from(url);
-
-        // Download remaining chunks using buffer_unordered for automatic batch processing
-        let chunk_results: Vec<u64> = stream::iter(0..total_chunks)
-            .map(|chunk_id| {
-                let start = resume_from + (chunk_id as u64 * chunk_size as u64);
-                let end = if chunk_id == total_chunks - 1 {
-                    total_size - 1 // Last chunk gets remainder
-                } else {
-                    start + chunk_size as u64 - 1
-                };
-
-                let chunk_path = temp_dir.join(format!("{}.{}.resume{}",
-                    &filename, download_id, chunk_id));
-
-                // Clone what we need for the async block (O(1) Arc clone)
-                let downloader = Self {
-                    client: self.client.clone(),
-                    config: Arc::clone(&self.config),
-                };
-                let url = Arc::clone(&url_shared);  // Cheap Arc clone (~5ns)
-                let progress_counter = Some(downloaded.clone());
-
-                async move {
-                    let result = downloader.download_range_with_progress(
-                        &url, start, end, &chunk_path, progress_counter
-                    ).await;
-                    if let Err(ref e) = result {
-                        error!("Chunk {chunk_id} failed: {e}");
-                    }
-                    result.map(|bytes| (chunk_id, bytes, chunk_path))
-                }
-            })
-            .buffer_unordered(self.config.concurrent_fragments) // Automatic batch processing!
-            .try_collect::<Vec<_>>()
-            .await
-            .inspect_err(|e| {
-                error!("Resume failed: {e}");
-                // Stop progress reporter on error
-                if let Some(task) = &progress_task {
-                    task.abort();
-                }
-                // Cleanup partial chunk files
-                let temp_dir = temp_dir.to_path_buf();
-                let filename = filename.clone();
-                tokio::spawn(async move {
-                    cleanup_chunk_files(&temp_dir, &filename, download_id, total_chunks).await;
-                });
-            })?
-            .into_iter()
-            .map(|(_, bytes, _)| bytes)
-            .collect();
-
-        // Sum up all downloaded bytes (just the new chunks)
-        let newly_downloaded: u64 = chunk_results.iter().sum();
-        info!("All {} chunks completed ({} MB new data)", total_chunks, newly_downloaded / 1024 / 1024);
-
-        // Stop progress reporter on success
-        if let Some(task) = progress_task {
-            task.abort();
-        }
-
-        // Append chunks to existing file in correct order
-        let file = tokio::fs::OpenOptions::new()
-            .append(true)
-            .open(path)
-            .await
-            .map_err(RdlpError::Io)?;
-        let mut writer = BufWriter::with_capacity(self.config.buffer_size, file);
-
-        info!("Appending {total_chunks} chunks to existing file...");
-        let mut deleted_chunks = 0;
-        for chunk_id in 0..total_chunks {
-            let chunk_path = temp_dir.join(format!("{}.{}.resume{}",
-                &filename, download_id, chunk_id));
-
-            let mut chunk_file = File::open(&chunk_path).await.map_err(RdlpError::Io)?;
-            tokio::io::copy(&mut chunk_file, &mut writer)
-                .await
-                .map_err(RdlpError::Io)?;
-
-            // Delete chunk file after appending
-            if tokio::fs::remove_file(&chunk_path).await.is_ok() {
-                deleted_chunks += 1;
-            }
-
-            if (chunk_id + 1) % 100 == 0 || chunk_id == total_chunks - 1 {
-                debug!("Appended {}/{} chunks", chunk_id + 1, total_chunks);
-            }
-        }
-        debug!("Cleaned up {deleted_chunks} chunk files");
-
-        writer.flush().await.map_err(RdlpError::Io)?;
-
-        let duration = start_time.elapsed();
-        let total_downloaded = resume_from + newly_downloaded;
-        // Fix Issue #2: Use 0 for retry_count (we don't track retries at this level)
-        let stats = DownloadStats::new(total_downloaded, duration, 0);
-
-        info!("Resume complete: {} MB total, {} MB new in {:.1}s ({:.1} MB/s)",
-            total_downloaded / 1024 / 1024,
-            newly_downloaded / 1024 / 1024,
-            duration.as_secs_f64(),
-            (newly_downloaded as f64 / duration.as_secs_f64()) / 1024.0 / 1024.0);
-
-        Ok(stats)
+impl Default for HttpDownloader {
+    fn default() -> Self {
+        Self::new()
     }
 }
 
@@ -719,39 +252,46 @@ impl Downloader for HttpDownloader {
         path: &Path,
         progress: Option<Box<dyn ProgressCallback>>,
     ) -> Result<DownloadStats> {
-        // Try to get file size and check for parallel support
         let size = self.get_size(url).await.ok().flatten();
         let supports_ranges = self.supports_ranges(url).await.unwrap_or(false);
 
-        debug!("Download analysis: size={} MB, concurrent={}, ranges={}",
+        debug!(
+            "Download analysis: size={} MB, concurrent={}, ranges={}",
             size.map(|s| s / 1024 / 1024).unwrap_or(0),
             self.config.concurrent_fragments,
-            supports_ranges);
+            supports_ranges
+        );
 
-        // If HEAD didn't return valid size, try a small Range request to get it
+        // Try Range request if HEAD didn't return size
         let size = if (size.is_none() || size == Some(0)) && supports_ranges {
             debug!("HEAD didn't return valid size, trying Range request...");
-            match retry_with_backoff(&self.config.retry_config, "HTTP GET (size check)", |_attempt| {
-                let client = self.client.clone();
-                let url = url.to_string();
-                async move {
-                    client
-                        .get(&url)
-                        .header("Range", "bytes=0-0")
-                        .send()
-                        .await
-                        .map_err(|e| RdlpError::Network(format!("Failed to check size: {e}")))
-                }
-            })
+            match retry_with_backoff(
+                &self.config.retry_config,
+                "HTTP GET (size check)",
+                |_attempt| {
+                    let client = self.client.clone();
+                    let url = url.to_string();
+                    async move {
+                        client
+                            .get(&url)
+                            .header("Range", "bytes=0-0")
+                            .send()
+                            .await
+                            .map_err(|e| RdlpError::Network(format!("Failed to check size: {e}")))
+                    }
+                },
+            )
             .await
             {
                 Ok(response) => {
-                    // Parse Content-Range: bytes 0-0/618618881
                     if let Some(content_range) = response.headers().get("content-range") {
                         if let Ok(range_str) = content_range.to_str() {
                             if let Some(total_str) = range_str.split('/').nth(1) {
                                 let detected_size = total_str.parse::<u64>().ok();
-                                debug!("Detected size from Range: {} MB", detected_size.map(|s| s / 1024 / 1024).unwrap_or(0));
+                                debug!(
+                                    "Detected size from Range: {} MB",
+                                    detected_size.map(|s| s / 1024 / 1024).unwrap_or(0)
+                                );
                                 detected_size
                             } else {
                                 None
@@ -769,7 +309,6 @@ impl Downloader for HttpDownloader {
             size
         };
 
-        // Check if we have a valid size and should use parallel
         let use_parallel = match size {
             Some(s) if s > 10 * 1024 * 1024 => {
                 self.config.concurrent_fragments > 1 && supports_ranges
@@ -778,10 +317,14 @@ impl Downloader for HttpDownloader {
         };
 
         if use_parallel {
-            info!("Using parallel download mode ({} connections)", self.config.concurrent_fragments);
-            return self.download_parallel(url, path, size.unwrap(), progress).await;
+            info!(
+                "Using parallel download mode ({} connections)",
+                self.config.concurrent_fragments
+            );
+            return self
+                .download_parallel(url, path, size.unwrap(), progress)
+                .await;
         } else {
-            // Use match with guards for cleaner Option handling (no unwrap)
             let reason = match size {
                 None | Some(0) => "could not detect file size",
                 Some(s) if s <= 10 * 1024 * 1024 => "file too small for parallel",
@@ -789,11 +332,13 @@ impl Downloader for HttpDownloader {
                 Some(_) if !supports_ranges => "server doesn't support ranges",
                 Some(_) => "unknown reason",
             };
-            warn!("Using sequential download - reason: {reason} (size: {:?} MB, fragments: {}, ranges: {supports_ranges})",
-                size.map(|s| s / 1024 / 1024), self.config.concurrent_fragments);
+            warn!(
+                "Using sequential download - reason: {reason} (size: {:?} MB, fragments: {}, ranges: {supports_ranges})",
+                size.map(|s| s / 1024 / 1024),
+                self.config.concurrent_fragments
+            );
         }
 
-        // Fallback to sequential download
         self.download_sequential(url, path, progress).await
     }
 
@@ -827,37 +372,36 @@ impl Downloader for HttpDownloader {
     ) -> Result<DownloadStats> {
         let start_time = Instant::now();
 
-        // Make request with Range header (with retry)
-        let response = retry_with_backoff(&self.config.retry_config, "HTTP GET (resume)", |_attempt| {
-            let client = self.client.clone();
-            let url = url.to_string();
-            async move {
-                let response = client
-                    .get(&url)
-                    .header("Range", format!("bytes={resume_from}-"))
-                    .send()
-                    .await
-                    .map_err(|e| RdlpError::Network(format!("Failed to fetch URL: {e}")))?;
+        let response = retry_with_backoff(
+            &self.config.retry_config,
+            "HTTP GET (resume)",
+            |_attempt| {
+                let client = self.client.clone();
+                let url = url.to_string();
+                async move {
+                    let response = client
+                        .get(&url)
+                        .header("Range", format!("bytes={resume_from}-"))
+                        .send()
+                        .await
+                        .map_err(|e| RdlpError::Network(format!("Failed to fetch URL: {e}")))?;
 
-                // Check if server supports range requests
-                if response.status().as_u16() != 206 {
-                    // Server doesn't support partial content - DO NOT overwrite partial file
-                    return Err(RdlpError::Download(format!(
-                        "Server does not support resume (expected HTTP 206, got {}). \
-                         Cannot continue download without overwriting existing data. \
-                         Please delete the partial file and restart the download.",
-                        response.status()
-                    )));
+                    if response.status().as_u16() != 206 {
+                        return Err(RdlpError::Download(format!(
+                            "Server does not support resume (expected HTTP 206, got {}). \
+                             Cannot continue download without overwriting existing data. \
+                             Please delete the partial file and restart the download.",
+                            response.status()
+                        )));
+                    }
+
+                    Ok(response)
                 }
-
-                Ok(response)
-            }
-        })
+            },
+        )
         .await?;
 
-        // Try to get total size from Content-Range header (more reliable for resume)
         let total_size = if let Some(content_range) = response.headers().get("content-range") {
-            // Content-Range: bytes 67108864-618618880/618618881
             if let Ok(range_str) = content_range.to_str() {
                 if let Some(total_str) = range_str.split('/').nth(1) {
                     total_str.parse::<u64>().ok()
@@ -868,50 +412,59 @@ impl Downloader for HttpDownloader {
                 None
             }
         } else {
-            // Fallback to Content-Length + resume_from
             response.content_length().map(|size| size + resume_from)
         };
 
-        // Check if we should use parallel resume
+        // Check for parallel resume
         if let Some(total) = total_size {
             let progress_pct = (resume_from as f64 / total as f64) * 100.0;
             let remaining_size = total - resume_from;
-            let supports_ranges = response.headers().get("accept-ranges")
+            let supports_ranges = response
+                .headers()
+                .get("accept-ranges")
                 .and_then(|v| v.to_str().ok())
                 .map(|v| v != "none")
-                .unwrap_or(true); // Assume true if we got 206
+                .unwrap_or(true);
 
-            debug!("Resume analysis: {:.1}% ({} MB / {} MB), remaining={} MB, concurrent={}, ranges={}",
-                progress_pct, resume_from / 1024 / 1024, total / 1024 / 1024,
+            debug!(
+                "Resume analysis: {:.1}% ({} MB / {} MB), remaining={} MB, concurrent={}, ranges={}",
+                progress_pct,
+                resume_from / 1024 / 1024,
+                total / 1024 / 1024,
                 remaining_size / 1024 / 1024,
                 self.config.concurrent_fragments,
-                supports_ranges);
+                supports_ranges
+            );
 
-            let can_parallel = remaining_size > 10 * 1024 * 1024 // > 10MB remaining for parallel to be worth it
+            let can_parallel = remaining_size > 10 * 1024 * 1024
                 && self.config.concurrent_fragments > 1
                 && supports_ranges;
 
-            // Use parallel resume if remaining size is large enough
             if can_parallel {
-                info!("Using parallel resume mode ({} connections), keeping {} MB, parallelizing {} MB",
+                info!(
+                    "Using parallel resume mode ({} connections), keeping {} MB, parallelizing {} MB",
                     self.config.concurrent_fragments,
-                    resume_from / 1024 / 1024, remaining_size / 1024 / 1024);
+                    resume_from / 1024 / 1024,
+                    remaining_size / 1024 / 1024
+                );
 
-                // Close the current response
                 drop(response);
-
-                // Resume with parallel download of remaining chunks
-                return self.download_parallel_resume(url, path, resume_from, total, progress).await;
+                return self
+                    .download_parallel_resume(url, path, resume_from, total, progress)
+                    .await;
             } else if !can_parallel {
-                warn!("Parallel resume not available (remaining: {} MB, concurrent: {}, ranges: {}), using sequential",
-                    remaining_size / 1024 / 1024, self.config.concurrent_fragments, supports_ranges);
+                warn!(
+                    "Parallel resume not available (remaining: {} MB, concurrent: {}, ranges: {}), using sequential",
+                    remaining_size / 1024 / 1024,
+                    self.config.concurrent_fragments,
+                    supports_ranges
+                );
             }
         }
 
         let content_length = response.content_length();
         let total_size = total_size.or_else(|| content_length.map(|size| size + resume_from));
 
-        // Open file in append mode with buffered writer
         let file = tokio::fs::OpenOptions::new()
             .append(true)
             .open(path)
@@ -919,7 +472,6 @@ impl Downloader for HttpDownloader {
             .map_err(RdlpError::Io)?;
         let mut writer = BufWriter::with_capacity(self.config.buffer_size, file);
 
-        // Download remaining data
         let mut stream = response.bytes_stream();
         let mut downloaded = resume_from;
         let mut last_update = Instant::now();
@@ -929,13 +481,9 @@ impl Downloader for HttpDownloader {
             let chunk = chunk_result
                 .map_err(|e| RdlpError::Network(format!("Failed to read chunk: {e}")))?;
 
-            writer.write_all(&chunk)
-                .await
-                .map_err(RdlpError::Io)?;
-
+            writer.write_all(&chunk).await.map_err(RdlpError::Io)?;
             downloaded += chunk.len() as u64;
 
-            // Update progress
             if let Some(ref callback) = progress {
                 let now = Instant::now();
                 if now.duration_since(last_update) >= update_interval {
